+++ replictl/src/commands/check/mod.rs
@@ -5,6 +5,7 @@ use clap::SubCommand;
 
 mod config;
 mod store;
+mod streams;
 
 use super::super::Interfaces;
 use super::super::Result;
@@ -22,6 +23,7 @@ pub fn command() -> App<'static, 'static> {
         .about("Perform checks on the system to find issues")
         .subcommand(config::command())
         .subcommand(store::command())
+        .subcommand(streams::command())
         .subcommand(SubCommand::with_name(DEEP_COMMAND)
             .about("Run all checks INCLUDING the ones that iterate over ALL data")
         )
@@ -41,6 +43,7 @@ pub fn run<'a>(args: &ArgMatches<'a>, interfaces: &Interfaces) -> Result<()> {
     match command {
         Some(config::COMMAND) => config::run(args, interfaces),
         Some(store::COMMAND) => store::run(args, interfaces),
+        Some(streams::COMMAND) => streams::run(args, interfaces),
         Some(DEEP_COMMAND) => run_deep(args, interfaces),
         Some(QUICK_COMMAND) => run_quick(args, interfaces),
         // Currently update is an alias for `deep` but that may change.
@@ -56,9 +59,11 @@ fn run_deep<'a>(args: &ArgMatches<'a>, interfaces: &Interfaces) -> Result<()> {
     let config = config::run(args, interfaces);
     let store_schema = store::schema(args, interfaces);
     let store_data = store::data(args, interfaces);
+    let streams_events = streams::events(args, interfaces);
     config?;
     store_schema?;
     store_data?;
+    streams_events?;
     Ok(())
 }